Move `withAccordionItemState` into an individual file
