+++ src/index.ts
@@ -1,7 +1,8 @@
 export { Accordion } from './components/Accordion';
-export { AccordionItem, withAccordionItemState } from './components/AccordionItem';
+export { AccordionItem } from './components/AccordionItem';
 export { AccordionProvider } from './components/AccordionProvider';
 export { ControlledAccordion } from './components/ControlledAccordion';
+export { withAccordionItemState } from './components/withAccordionItemState';
 export { useAccordion } from './hooks/useAccordion';
 export { useAccordionItem } from './hooks/useAccordionItem';
 export { useAccordionItemState } from './hooks/useAccordionItemState';
@@ -10,8 +11,9 @@ export { useAccordionState } from './hooks/useAccordionState';
 export { useHeightTransition } from './hooks/useHeightTransition';
 export { useMergeRef } from './hooks/useMergeRef';
 export type { AccordionProps } from './components/Accordion';
-export type { AccordionItemProps, ItemStateProps, ItemModifiers } from './components/AccordionItem';
+export type { AccordionItemProps, ItemModifiers } from './components/AccordionItem';
 export type { ControlledAccordionProps } from './components/ControlledAccordion';
+export type { ItemStateProps } from './components/withAccordionItemState';
 export type {
   AccordionProviderOptions,
   AccordionProviderValue,
