+++ types/components/AccordionItem.d.ts
@@ -1,4 +1,4 @@
-import { ReactNode, RefObject, ForwardedRef, MemoExoticComponent } from 'react';
+import { ReactNode, ForwardedRef } from 'react';
 import { TransitionStatus } from 'react-transition-state';
 import { ElementProps, ItemState, ItemStateOptions } from '../utils/constants';
 declare type ItemModifiers = {
@@ -17,10 +17,5 @@ interface AccordionItemProps extends ItemStateOptions, ElementProps<HTMLDivEleme
     contentProps?: ItemElementProps<HTMLDivElement>;
     panelProps?: ItemElementProps<HTMLDivElement>;
 }
-interface ItemStateProps<E extends Element, T = E> extends ItemState {
-    itemRef: RefObject<E>;
-    forwardedRef: ForwardedRef<T>;
-}
-declare const withAccordionItemState: <P extends ItemStateOptions, E extends Element, T = E>(WrappedItem: MemoExoticComponent<(props: ItemStateProps<E, T>) => JSX.Element>) => import("react").ForwardRefExoticComponent<import("react").PropsWithoutRef<P> & import("react").RefAttributes<T>>;
 declare const AccordionItem: import("react").ForwardRefExoticComponent<AccordionItemProps & import("react").RefAttributes<HTMLDivElement>>;
-export { withAccordionItemState, AccordionItem, AccordionItemProps, ItemStateProps, ItemModifiers };
+export { AccordionItem, AccordionItemProps, ItemModifiers };
